Articles can now parse their own headlines to create words and mentions
Parsing creates a hash that stores words and the number of times those words appear in
the headline as a key:value pair. Words will only be created in the database if they
don't already exist.  Mentions are linked to words and articles...parsing is still dumb
and includes punctuation, etc...